backend/requirements.txt
uvicorn
fastapi
python-multipart
deepgram-sdk
openai==0.28